Rename MenuDropdown component to MenuItemsDropdown to finally avoid any case mismatches
