+++ plugins/CoreHome/vue/src/MenuItemsDropdown/MenuItemsDropdown.less
