+++ plugins/CoreHome/vue/src/MenuItemsDropdown/MenuItemsDropdown.vue
