+++ plugins/CoreHome/vue/src/index.ts
@@ -30,7 +30,7 @@ import './MatomoDialog/MatomoDialog.adapter';
 import './EnrichedHeadline/EnrichedHeadline.adapter';
 import './ContentBlock/ContentBlock.adapter';
 import './Comparisons/Comparisons.adapter';
-import './MenuDropdown/MenuDropdown.adapter';
+import './MenuItemsDropdown/MenuItemsDropdown.adapter';
 import './DatePicker/DatePicker.adapter';
 import './DateRangePicker/DateRangePicker.adapter';
 import './PeriodDatePicker/PeriodDatePicker.adapter';
@@ -92,7 +92,7 @@ export { default as SideNav } from './SideNav/SideNav';
 export { default as EnrichedHeadline } from './EnrichedHeadline/EnrichedHeadline.vue';
 export { default as ContentBlock } from './ContentBlock/ContentBlock.vue';
 export { default as Comparisons } from './Comparisons/Comparisons.vue';
-export { default as MenuDropdown } from './MenuDropdown/MenuDropdown.vue';
+export { default as MenuItemsDropdown } from './MenuItemsDropdown/MenuItemsDropdown.vue';
 export { default as DatePicker } from './DatePicker/DatePicker.vue';
 export { default as DateRangePicker } from './DateRangePicker/DateRangePicker.vue';
 export { default as PeriodDatePicker } from './PeriodDatePicker/PeriodDatePicker.vue';
