+++ plugins/Morpheus/stylesheets/base.less
@@ -14,7 +14,7 @@
 @import "uibase/_header.less";
 @import "uibase/_headerMessage.less";
 @import "../../CoreHome/vue/src/SiteSelector/SiteSelector.less";
-@import "../../CoreHome/vue/src/MenuDropdown/MenuDropdown.less";
+@import "../../CoreHome/vue/src/MenuItemsDropdown/MenuItemsDropdown.less";
 @import "../../CoreHome/vue/src/Alert/alert.less";
 @import "../../CoreHome/stylesheets/dataTable/_entityTable.less";
 @import "uibase/_periodSelect.less";
